# fix: batch insert logs