--- a/server/internal/db/sqlite_test.go
+++ b/server/internal/db/sqlite_test.go
@@ -58,14 +58,14 @@ func TestSqliteProvider(t *testing.T) {
 		require.NotEmpty(tc, sessionID)
 		require.NotNil(tc, session)
 
-		err = session.WriteLogLine(ctx, WriteLogLinePayload{
+		err = session.WriteLogLinesBatch(ctx, WriteLogLinesBatchPayload{
 			Timestamp: time.Now(),
-			Line:      "hello",
+			Lines:     []string{"hello"},
 		})
 		require.NoError(tc, err)
-		err = session.WriteLogLine(ctx, WriteLogLinePayload{
+		err = session.WriteLogLinesBatch(ctx, WriteLogLinesBatchPayload{
 			Timestamp: time.Now(),
-			Line:      "world",
+			Lines:     []string{"world"},
 		})
 		require.NoError(tc, err)
 		sqliteSession := session.(*SqliteSession)