fix #43 

Before:

```
goos: darwin
goarch: arm64
pkg: github.com/b4fun/ku/server/internal/db
BenchmarkSessionLogWriter-10    	       3	11708229417 ns/op
PASS
ok  	github.com/b4fun/ku/server/internal/db	44.459s
```

After:

```
goos: darwin
goarch: arm64
pkg: github.com/b4fun/ku/server/internal/db
BenchmarkSessionLogWriter-10    	    1021	  35460912 ns/op
PASS
ok  	github.com/b4fun/ku/server/internal/db	41.209s
```